Finish UI Architecture refactor. Closes #78.
