+++ client/src/components/AuthForms/SignUp.js
@@ -1,65 +1,79 @@
-import React from 'react';
-import { NavLink, useHistory } from 'react-router-dom';
+import React from "react";
+import { NavLink, useHistory } from "react-router-dom";
 import { connect } from 'react-redux';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
 
 
-import google from '../images/google.png';
-import sprite from '../images/sprite.svg';
-import { handleLocalSignin } from '../actions/auth'; 
+import google from "../../images/google.png";
+import sprite from "../../images/sprite.svg";
+import { handleLocalSignup } from '../../actions/auth';
 
-function SignIn(props) {
+function SignUp(props) {
   const history = useHistory();
   const formik = useFormik({
     initialValues: {
+      name: '',
       email: '',
       password: ''
     },
     onSubmit: values => {
-      const authArgs = Object.assign({}, values, { history });
-      props.handleLocalSignin(authArgs)
+      const signUpArgs = Object.assign({}, values, { history });
+      props.handleLocalSignup(signUpArgs);
     },
-    validationSchema: signinSchema
+    validationSchema: signupSchema
   });
 
   return (
     <>
-      <div className="mt-12 max-w-lg m-auto bg-white shadow-lg rounded px-8 py-10 border">
-        <h2 className="text-center mb-6 font-semibold text-3xl">Sign in</h2>
+      <div className="mt-10 max-w-lg m-auto bg-white shadow-lg rounded px-8 py-10 border">
+        <h2 className="text-center mb-6 font-semibold text-3xl">Sign up</h2>
         <div className="flex flex-col">
-          <a
-            className=" border rounded flex items-center shadow px-4 py-3 bg-white flex-grow justify-center focus:outline-none"
-            onClick={props.handleGoogleAuth}
-            href="/auth/google"
-          >
+          <button className=" border rounded flex items-center shadow px-4 py-3 bg-white flex-grow justify-center focus:outline-none">
             <img src={google} alt="google" className="h-8 w-8" />
-            <span className="ml-4 ">Sign in with Google</span>
-          </a>
-          <a
-            className=" border rounded flex items-center shadow mt-6 px-4 py-3 bg-white justify-center focus:outline-none"
-            href="/auth/github"
-          >
+            <span className="ml-4">Sign up with Google</span>
+          </button>
+          <button className=" border rounded flex items-center shadow mt-6 px-4 py-3 bg-white justify-center focus:outline-none">
             <svg className="h-8 w-8">
               <use xlinkHref={`${sprite}#icon-github`} />
             </svg>
-            <span className="ml-4">Sign in with GitHub</span>
-          </a>
+            <span className="ml-4 ">Sign up with GitHub</span>
+          </button>
         </div>
 
         <p className="text-center my-8 text-lg font-semibold">
-          Or Sign in with Email
+          Or Sign up with Email
         </p>
 
         <form className="flex flex-col" onSubmit={formik.handleSubmit}>
-          {
-            props.error && (
-              <div className="text-red-500 text-center font-semibold">
-                {props.error}
-              </div>
-            )
-          }
+          {props.error && (
+            <div className="text-red-500 text-center font-semibold">
+              {props.error}
+            </div>
+          )}
           <div className="flex flex-col">
+            <label
+              htmlFor="name"
+              className="ml-3 font-medium uppercase text-sm text-gray-700"
+            >
+              Name
+            </label>
+            <input
+              type="text"
+              id="name"
+              name="name"
+              className="mt-2 py-3 px-4 rounded bg-gray-200 border"
+              onChange={formik.handleChange}
+              value={formik.values.name}
+            />
+            {formik.errors.name && formik.touched.name ? (
+              <div className="text-red-500 text-sm px-4">
+                {formik.errors.name}
+              </div>
+            ) : null}
+          </div>
+
+          <div className="flex flex-col mt-6">
             <label
               htmlFor="email"
               className="ml-3 font-medium uppercase text-sm text-gray-700"
@@ -67,7 +81,7 @@ function SignIn(props) {
               Email
             </label>
             <input
-              type="email"
+              type="text"
               id="email"
               name="email"
               className="mt-2 py-3 px-4 rounded bg-gray-200 border"
@@ -107,25 +121,26 @@ function SignIn(props) {
             className="shadow rounded px-4 py-3 bg-indigo-600 text-white mt-10 uppercase text-sm"
             type="submit"
           >
-            Sign in
+            Sign up
           </button>
         </form>
       </div>
 
       <div className="mt-10 max-w-lg m-auto bg-gray-300 rounded text-center py-4">
-        Don't have an account?
+        Already have an account?
         <NavLink
-          to="/index/signup"
+          to="/index/signin"
           className="ml-3 font-semibold hover:underline"
         >
-          Sign up
+          Sign in
         </NavLink>
       </div>
     </>
   );
 }
 
-const signinSchema = Yup.object().shape({
+const signupSchema = Yup.object().shape({
+  name: Yup.string().required("Name is required"),
   email: Yup.string().email().required("Email is required"),
   password: Yup.string()
     .matches(
@@ -137,8 +152,9 @@ const signinSchema = Yup.object().shape({
     .required("Password is required"),
 });
 
-const mapStateToProps = state => ({ error: state.error });
+
+const mapStateToProps = state => ({ error: state.error })
 export default connect(
   mapStateToProps,
-  { handleLocalSignin }
-)(SignIn);
+  { handleLocalSignup }
+)(SignUp);
