+++ client/src/components/Issue/ManageIssues.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 
 
-import sprite from "./images/sprite.svg";
+import sprite from "../../images/sprite.svg";
 
 export default function ManageIssues({ url }) {
